[IMP] web_editor, *: move the left panel to the right
*: website, mass_mailing

See parent commit for information about the scrollbar moved to the
'#wrapwrap' element, which was necessary to achieve this right panel
design. Note: in the future, having the scroll on the #wrapwrap may
not be necessary anymore as the edition might be done in an iframe.

Part of https://github.com/odoo/odoo/pull/57975
task-2344227

X-original-commit: 2abe8517063b2696cc89e9a453dcc3921fb3b8c7