[FIX] web_editor, website: fix menu colors in table of content
Before this commit, the colors of the navbar in the table of content
menu were not adapted to the color combinations system. With navbar in
top position, the background and the colors of the links were using the
same color.

After this commit, the colors of the list-group-item links are also
adapted with color combinations.

task-2312878

closes odoo/odoo#58645

X-original-commit: b30255bf422d2362de7917221c00b86b34c616e0
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>