[FIX] website: correctly add o_top_fixed_element class on header
The class is supposed to be placed on elements which are absolutely
positioned over the main scrolling element at the top of the page. It
was wrongly placed on headers as soon as they become fixed while they
should have been placed on headers as soon as they become fixed *and*
over the main scrolling element.

Part of https://github.com/odoo/odoo/pull/57975
task-2344227

X-original-commit: ee47bd9ce0da526ea5b3387a354ab8fbd06cc55f