[FIX] website_form: layout issue on multi checkboxes options
Before this commit, buttons to remove checkboxes and buttons to
re-order checkboxes were styled like button.

After this commit, buttons are just icons. We also improve a little
bit the design of the checkbox list (add space between elements and a
light border between items in the checkbox list).

task-2312878

closes odoo/odoo#58181

X-original-commit: 0253b01632a30adfe071aa090a875fb3915e1962
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>