[FIX] website_slides: hide the website default header in fullscreen
Bug
===
You can enter in a fullscreen mode in the frontend, of the slide channel.
In this view, the default website header should not be visible.

To do that, we play with z-index, but since 5152fb12e15ce9c6f66505a0f67dad6b3bfd3035 ,
the z-index of the website header has been changed without changing the z-index
of the fullscreen view.

Task-2345016

closes odoo/odoo#58483

X-original-commit: 8d35540df6d21af6fe716f73aefa37bfb66c7ac2
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>