[FIX] website: restore top editor bar for translate mode
The text tools were moved in the right panel with a temporary solution
but as there is no right panel in translate mode, the text tools were
just floating summernote-way. This commit restores the top editor bar
in that case while waiting for the new editor.

Part of https://github.com/odoo/odoo/pull/58347

closes odoo/odoo#58372

X-original-commit: b9d4ccfc8db1f3425b8ee0e8d6311ebaac0f26b2
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>